# backend/docs/openapi.json
{
  "components": {
    "schemas": {
...
      "CompetencyCreate": {
        "additionalProperties": false,
        "properties": {
          "active": {
            "default": true,
            "title": "Active",
            "type": "boolean"
          },
          "category": {
            "maxLength": 80,
            "minLength": 2,
            "title": "Category",
            "type": "string"
          },
          "description": {
            "maxLength": 2000,
            "minLength": 3,
            "title": "Description",
            "type": "string"
          },
          "name": {
            "maxLength": 160,
            "minLength": 2,
            "title": "Name",
            "type": "string"
          },
          "prerequisite_slugs": {
            "items": {
              "type": "string"
            },
            "maxItems": 30,
            "title": "Prerequisite Slugs",
            "type": "array"
          },
          "slug": {
            "maxLength": 100,
            "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$",
            "title": "Slug",
            "type": "string"
          }
        },
        "required": [
          "slug",
          "name",
          "description",
          "category"
        ],
        "title": "CompetencyCreate",
        "type": "object"
      },
...
      "CompetencyRead": {
        "additionalProperties": false,
        "properties": {
          "active": {
            "default": true,
            "title": "Active",
            "type": "boolean"
          },
          "category": {
            "maxLength": 80,
            "minLength": 2,
            "title": "Category",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "maxLength": 2000,
            "minLength": 3,
            "title": "Description",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "maxLength": 160,
            "minLength": 2,
            "title": "Name",
            "type": "string"
          },
          "prerequisite_slugs": {
            "items": {
              "type": "string"
            },
            "maxItems": 30,
            "title": "Prerequisite Slugs",
            "type": "array"
          },
          "slug": {
            "maxLength": 100,
            "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$",
            "title": "Slug",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "slug",
          "name",
          "description",
          "category",
          "id",
          "version",
          "created_at",
          "updated_at"
        ],
        "title": "CompetencyRead",
        "type": "object"
      },
      "CompetencyUpdate": {
        "additionalProperties": false,
        "properties": {
          "active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active"
          },
          "category": {
            "anyOf": [
              {
                "maxLength": 80,
                "minLength": 2,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "description": {
            "anyOf": [
              {
                "maxLength": 2000,
                "minLength": 3,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 160,
                "minLength": 2,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "prerequisite_slugs": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "maxItems": 30,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prerequisite Slugs"
          }
        },
        "title": "CompetencyUpdate",
        "type": "object"
      },
...
      "FeasibilityCategory": {
        "enum": [
          "feasible",
          "feasible_with_constraints",
          "challenging_but_possible",
          "unlikely_under_current_conditions",
          "insufficient_information",
          "requires_human_review"
        ],
        "title": "FeasibilityCategory",
        "type": "string"
      },
      "FeasibilityRequest": {
        "additionalProperties": false,
        "properties": {
          "excluded_competencies": {
            "items": {
              "type": "string"
            },
            "maxItems": 30,
            "title": "Excluded Competencies",
            "type": "array"
          },
          "existing_commitment_hours_per_week": {
            "default": 0,
            "maximum": 80.0,
            "minimum": 0.0,
            "title": "Existing Commitment Hours Per Week",
            "type": "number"
          },
          "target_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Date"
          },
          "weekly_hours": {
            "anyOf": [
              {
                "maximum": 80.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weekly Hours"
          }
        },
        "title": "FeasibilityRequest",
        "type": "object"
      },
      "FeasibilityResult": {
        "additionalProperties": false,
        "properties": {
          "assumptions": {
            "items": {
              "type": "string"
            },
            "title": "Assumptions",
            "type": "array"
          },
          "available_hours": {
            "minimum": 0.0,
            "title": "Available Hours",
            "type": "number"
          },
          "category": {
            "$ref": "#/components/schemas/FeasibilityCategory"
          },
          "confidence": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "type": "number"
          },
          "decision_cards": {
            "items": {
              "$ref": "#/components/schemas/DecisionCardData"
            },
            "title": "Decision Cards",
            "type": "array"
          },
          "estimated_effort_hours": {
            "$ref": "#/components/schemas/EffortRange"
          },
          "major_risks": {
            "items": {
              "type": "string"
            },
            "title": "Major Risks",
            "type": "array"
          },
          "recommended_adjustments": {
            "items": {
              "type": "string"
            },
            "title": "Recommended Adjustments",
            "type": "array"
          },
          "scenario_options": {
            "items": {
              "type": "string"
            },
            "title": "Scenario Options",
            "type": "array"
          },
          "weekly_capacity_hours": {
            "minimum": 0.0,
            "title": "Weekly Capacity Hours",
            "type": "number"
          },
          "weeks_available": {
            "minimum": 0.0,
            "title": "Weeks Available",
            "type": "number"
          }
        },
        "required": [
          "category",
          "estimated_effort_hours",
          "available_hours",
          "weekly_capacity_hours",
          "weeks_available",
          "major_risks",
          "recommended_adjustments",
          "scenario_options",
          "confidence",
          "assumptions",
          "decision_cards"
        ],
        "title": "FeasibilityResult",
        "type": "object"
      },
...
      "GoalClarificationRequest": {
        "additionalProperties": false,
        "properties": {
          "constraints": {
            "items": {
              "type": "string"
            },
            "maxItems": 20,
            "title": "Constraints",
            "type": "array"
          },
          "desired_outcome": {
            "anyOf": [
              {
                "maxLength": 500,
                "minLength": 3,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Desired Outcome"
          },
          "raw_goal": {
            "anyOf": [
              {
                "maxLength": 4000,
                "minLength": 5,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Goal"
          },
          "target_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Date"
          },
          "target_level": {
            "anyOf": [
              {
                "maximum": 5.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Level"
          },
          "template_slug": {
            "anyOf": [
              {
                "maxLength": 80,
                "minLength": 2,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Template Slug"
          },
          "weekly_hours": {
            "anyOf": [
              {
                "maximum": 80.0,
                "minimum": 0.5,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weekly Hours"
          }
        },
        "title": "GoalClarificationRequest",
        "type": "object"
      },
...
      "GoalClarificationResult": {
        "additionalProperties": false,
        "properties": {
          "clarified_goal": {
            "$ref": "#/components/schemas/ClarifiedGoal"
          },
          "decision_cards": {
            "items": {
              "$ref": "#/components/schemas/DecisionCardData"
            },
            "title": "Decision Cards",
            "type": "array"
          }
        },
        "required": [
          "clarified_goal",
          "decision_cards"
        ],
        "title": "GoalClarificationResult",
        "type": "object"
      },
      "GoalCompetenciesResult": {
        "additionalProperties": false,
        "properties": {
          "competencies": {
            "items": {
              "$ref": "#/components/schemas/CompetencyGap"
            },
            "title": "Competencies",
            "type": "array"
          },
          "goal_id": {
            "format": "uuid",
            "title": "Goal Id",
            "type": "string"
          },
          "template_slug": {
            "title": "Template Slug",
            "type": "string"
          }
        },
        "required": [
          "goal_id",
          "template_slug",
          "competencies"
        ],
        "title": "GoalCompetenciesResult",
        "type": "object"
      },
...
      "GoalGraphResult": {
        "additionalProperties": false,
        "properties": {
          "core_path": {
            "items": {
              "type": "string"
            },
            "title": "Core Path",
            "type": "array"
          },
          "decision_cards": {
            "items": {
              "$ref": "#/components/schemas/DecisionCardData"
            },
            "title": "Decision Cards",
            "type": "array"
          },
          "edges": {
            "items": {
              "$ref": "#/components/schemas/GraphEdgeRead"
            },
            "title": "Edges",
            "type": "array"
          },
          "estimated_duration_weeks": {
            "title": "Estimated Duration Weeks",
            "type": "number"
          },
          "goal_id": {
            "format": "uuid",
            "title": "Goal Id",
            "type": "string"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/GraphNodeRead"
            },
            "title": "Nodes",
            "type": "array"
          },
          "optional_branches": {
            "items": {
              "type": "string"
            },
            "title": "Optional Branches",
            "type": "array"
          }
        },
        "required": [
          "goal_id",
          "nodes",
          "edges",
          "core_path",
          "optional_branches",
          "estimated_duration_weeks",
          "decision_cards"
        ],
        "title": "GoalGraphResult",
        "type": "object"
      },
...
      "GoalTemplateCreate": {
        "additionalProperties": false,
        "properties": {
          "active": {
            "default": true,
            "title": "Active",
            "type": "boolean"
          },
          "category": {
            "maxLength": 80,
            "minLength": 2,
            "title": "Category",
            "type": "string"
          },
          "default_duration_weeks": {
            "maximum": 260.0,
            "minimum": 1.0,
            "title": "Default Duration Weeks",
            "type": "integer"
          },
          "default_target_level": {
            "default": 3,
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Default Target Level",
            "type": "integer"
          },
          "description": {
            "maxLength": 2000,
            "minLength": 3,
            "title": "Description",
            "type": "string"
          },
          "matching_terms": {
            "items": {
              "type": "string"
            },
            "maxItems": 50,
            "minItems": 1,
            "title": "Matching Terms",
            "type": "array"
          },
          "measurable_outcome": {
            "maxLength": 500,
            "minLength": 5,
            "title": "Measurable Outcome",
            "type": "string"
          },
          "name": {
            "maxLength": 160,
            "minLength": 3,
            "title": "Name",
            "type": "string"
          },
          "requirements": {
            "items": {
              "$ref": "#/components/schemas/TemplateRequirement"
            },
            "maxItems": 100,
            "minItems": 1,
            "title": "Requirements",
            "type": "array"
          },
          "slug": {
            "maxLength": 100,
            "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$",
            "title": "Slug",
            "type": "string"
          },
          "success_criteria": {
            "items": {
              "type": "string"
            },
            "maxItems": 20,
            "minItems": 1,
            "title": "Success Criteria",
            "type": "array"
          }
        },
        "required": [
          "slug",
          "name",
          "description",
          "category",
          "matching_terms",
          "default_duration_weeks",
          "measurable_outcome",
          "success_criteria",
          "requirements"
        ],
        "title": "GoalTemplateCreate",
        "type": "object"
      },
      "GoalTemplateRead": {
        "additionalProperties": false,
        "properties": {
          "active": {
            "default": true,
            "title": "Active",
            "type": "boolean"
          },
          "category": {
            "maxLength": 80,
            "minLength": 2,
            "title": "Category",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "default_duration_weeks": {
            "maximum": 260.0,
            "minimum": 1.0,
            "title": "Default Duration Weeks",
            "type": "integer"
          },
          "default_target_level": {
            "default": 3,
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Default Target Level",
            "type": "integer"
          },
          "description": {
            "maxLength": 2000,
            "minLength": 3,
            "title": "Description",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "matching_terms": {
            "items": {
              "type": "string"
            },
            "maxItems": 50,
            "minItems": 1,
            "title": "Matching Terms",
            "type": "array"
          },
          "measurable_outcome": {
            "maxLength": 500,
            "minLength": 5,
            "title": "Measurable Outcome",
            "type": "string"
          },
          "name": {
            "maxLength": 160,
            "minLength": 3,
            "title": "Name",
            "type": "string"
          },
          "requirements": {
            "items": {
              "$ref": "#/components/schemas/TemplateRequirement"
            },
            "maxItems": 100,
            "minItems": 1,
            "title": "Requirements",
            "type": "array"
          },
          "slug": {
            "maxLength": 100,
            "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$",
            "title": "Slug",
            "type": "string"
          },
          "success_criteria": {
            "items": {
              "type": "string"
            },
            "maxItems": 20,
            "minItems": 1,
            "title": "Success Criteria",
            "type": "array"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "slug",
          "name",
          "description",
          "category",
          "matching_terms",
          "default_duration_weeks",
          "measurable_outcome",
          "success_criteria",
          "requirements",
          "id",
          "version",
          "created_at",
          "updated_at"
        ],
        "title": "GoalTemplateRead",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SkillGapRequest": {
        "additionalProperties": false,
        "properties": {
          "competency_evidence": {
            "items": {
              "$ref": "#/components/schemas/StudentCompetencyInput"
            },
            "maxItems": 100,
            "title": "Competency Evidence",
            "type": "array"
          }
        },
        "title": "SkillGapRequest",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "AstraPath Phase 1-5 API. Contracts are versioned under /api/v1 and support exactly two roles: student and admin.",
    "title": "AstraPath API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/admin/competencies": {
      "post": {
        "operationId": "create_competency_api_v1_admin_competencies_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompetencyCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompetencyRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Competency",
        "tags": [
          "phase2-admin"
        ]
      }
    },
    "/api/v1/admin/competencies/{competency_id}": {
      "patch": {
        "operationId": "update_competency_api_v1_admin_competencies__competency_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "competency_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Competency Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompetencyUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompetencyRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Competency",
        "tags": [
          "phase2-admin"
        ]
      }
    },
    "/api/v1/admin/goal-templates": {
      "post": {
        "operationId": "create_goal_template_api_v1_admin_goal_templates_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoalTemplateCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoalTemplateRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Goal Template",
        "tags": [
          "phase2-admin"
        ]
      }
    },
...
    "/api/v1/goals/{goal_id}/clarify": {
      "post": {
        "operationId": "clarify_goal_api_v1_goals__goal_id__clarify_post",
        "parameters": [
          {
            "in": "path",
            "name": "goal_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Goal Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoalClarificationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoalClarificationResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Clarify Goal",
        "tags": [
          "goal-intelligence"
        ]
      }
    },
    "/api/v1/goals/{goal_id}/competencies": {
      "get": {
        "operationId": "get_goal_competencies_api_v1_goals__goal_id__competencies_get",
        "parameters": [
          {
            "in": "path",
            "name": "goal_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Goal Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoalCompetenciesResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Goal Competencies",
        "tags": [
          "goal-intelligence"
        ]
      }
    },
    "/api/v1/goals/{goal_id}/feasibility": {
      "post": {
        "operationId": "assess_feasibility_api_v1_goals__goal_id__feasibility_post",
        "parameters": [
          {
            "in": "path",
            "name": "goal_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Goal Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/FeasibilityRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeasibilityResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Assess Feasibility",
        "tags": [
          "goal-intelligence"
        ]
      }
    },
    "/api/v1/goals/{goal_id}/graph": {
      "get": {
        "operationId": "get_goal_graph_api_v1_goals__goal_id__graph_get",
        "parameters": [
          {
            "in": "path",
            "name": "goal_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Goal Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoalGraphResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Goal Graph",
        "tags": [
          "goal-intelligence"
        ]
      }
    },
...
    "/api/v1/goals/{goal_id}/skill-gap": {
      "post": {
        "operationId": "analyze_skill_gap_api_v1_goals__goal_id__skill_gap_post",
        "parameters": [
          {
            "in": "path",
            "name": "goal_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Goal Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/SkillGapRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillGapWorkflowResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Analyze Skill Gap",
        "tags": [
          "goal-intelligence"
        ]
      }
    },
...
  }
}